Swap out to async callback to allow more flexibility
The original code required a return from the `facetMatches` and
`valueMatches`.  This is less than ideal in an environment where you
might not know all of the values upfront.  For example, you might wish
to inspect the previously selected facets and provide a different set to
allow context-specific faceting.

Likewise, you might not be able to provide all of the values that could
match or might want to delegate the generating of a filtered list to a
search index in the background.