Fixing dash issue in facet names
- Facet names now can include dashes. Before this fix when facet names
are selected in autocomplete, facet name was being added as query text.
- Demo.html edited to show that label-value pair is available to use -
When "text:" is typed, an error was thrown due to wrong call of this
options and wrong usage of this.app.options.callbacks.facetMatches()
function - When facet name and column entered, a facet search is being
made. Now that includes not only the labels as well as values.